+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Version History
 
 -   Fix artifacts when sampling AMR volume
 -   Fix enum size on Windows with older MSVC
+-   Fix benchmark compilation with MSVC
 -   Update to latest versions of dependencies
 
 ### Open VKL 2.0.2